AudioT formats support.
They are used in Rise of the Triad and Wolf3D engine games.
AudioT PC speaker sounds have playback now (it was too similar to DMX PC
speaker sounds not to add).
AudioT Adlib sounds are identified.
Reduced reliability of PSX graphics identification, because one of the
Wolf sounds was misidentified as one.